# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CPNS & School Tutoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages/{packageId}/quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes for a package
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz (admin/tutor)
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz detail with questions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update quiz
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete quiz
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/questions:
    post:
      operationId: addQuizQuestion
      tags: [quizzes]
      summary: Add question to quiz
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"

  # Tryouts
  /tryouts:
    get:
      operationId: listTryouts
      tags: [tryouts]
      summary: List tryouts
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tryouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tryout"
    post:
      operationId: createTryout
      tags: [tryouts]
      summary: Create tryout (admin/tutor)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TryoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tryout"
  /tryouts/{id}:
    get:
      operationId: getTryout
      tags: [tryouts]
      summary: Get tryout detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tryout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TryoutDetail"
    patch:
      operationId: updateTryout
      tags: [tryouts]
      summary: Update tryout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TryoutUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tryout"
    delete:
      operationId: deleteTryout
      tags: [tryouts]
      summary: Delete tryout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tryouts/{id}/questions:
    post:
      operationId: addTryoutQuestion
      tags: [tryouts]
      summary: Add question to tryout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TryoutQuestionInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TryoutDetail"

  # Attempts (quiz & tryout)
...
  /tryouts/{id}/ranking:
    get:
      operationId: getTryoutRanking
      tags: [tryouts]
      summary: Get tryout national ranking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingEntry"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Question:
      type: object
      required: [id, questionText, optionA, optionB, optionC, optionD, optionE, correctAnswer, category, difficulty, createdAt]
      properties:
        id:
          type: integer
        questionText:
          type: string
        imageUrl:
          type: ["string", "null"]
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        optionE:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D, E]
        explanation:
          type: ["string", "null"]
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Quiz:
      type: object
      required: [id, packageId, title, timeLimit, passingScore, createdAt]
      properties:
        id:
          type: integer
        packageId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        timeLimit:
          type: integer
        passingScore:
          type: integer
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    QuizInput:
      type: object
      required: [title, timeLimit, passingScore]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        timeLimit:
          type: integer
        passingScore:
          type: integer
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        timeLimit:
          type: integer
        passingScore:
          type: integer
    QuizDetail:
      type: object
      required: [id, packageId, title, timeLimit, passingScore, questions, createdAt]
      properties:
        id:
          type: integer
        packageId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        timeLimit:
          type: integer
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
    QuizQuestionInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        orderIndex:
          type: ["integer", "null"]

    # Tryout
    Tryout:
      type: object
      required: [id, title, type, durationMinutes, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [CPNS_SKD, UTBK, CUSTOM]
        durationMinutes:
          type: integer
        scheduledAt:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    TryoutInput:
      type: object
      required: [title, type, durationMinutes]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [CPNS_SKD, UTBK, CUSTOM]
        durationMinutes:
          type: integer
        scheduledAt:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
    TryoutUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        durationMinutes:
          type: integer
        scheduledAt:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
    TryoutDetail:
      type: object
      required: [id, title, type, durationMinutes, questions, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        durationMinutes:
          type: integer
        scheduledAt:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/TryoutQuestion"
        createdAt:
          type: string
    TryoutQuestion:
      type: object
      required: [id, questionText, optionA, optionB, optionC, optionD, optionE, category]
      properties:
        id:
          type: integer
        questionText:
          type: string
        imageUrl:
          type: ["string", "null"]
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        optionE:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
    TryoutQuestionInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        subCategory:
          type: ["string", "null"]

    # Attempts
...
    RankingEntry:
      type: object
      required: [rank, userId, name, score]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        score:
          type: number
        isCurrentUser:
          type: boolean