# lib/api-zod/src/generated/types/tryoutQuestionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPNS & School Tutoring Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TryoutQuestionInput {
  questionId: number;
  /** @nullable */
  subCategory?: string | null;
}